# docs/v3/openapi.yaml
openapi: 3.0.1
info:
  contact: {}
  description: REST API for managing software licenses
  title: License Server API
  version: 0.1.0
servers:
- url: http://localhost:8080/api/v1
paths:
...
  /admin/users:
    get:
      description: retrieves all users
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/httputil.HTTPResponse-array_model_User'
          description: OK
        "500":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/httputil.HTTPError'
          description: Internal Server Error
      security:
      - ApiKey: []
      summary: Get all users
      tags:
      - users
    post:
      description: creates a new user
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/model.User'
        description: new user info
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/httputil.HTTPResponse-model_User'
          description: OK
        "400":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/httputil.HTTPError'
          description: Bad Request
        "500":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/httputil.HTTPError'
          description: Internal Server Error
      security:
      - ApiKey: []
      summary: Add a user
      tags:
      - users
      x-codegen-request-body-name: data
  /admin/users/:id:
    delete:
      description: deletes a specific user
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/httputil.HTTPResponse-model_User'
          description: OK
        "400":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/httputil.HTTPError'
          description: Bad Request
        "500":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/httputil.HTTPError'
          description: Internal Server Error
      security:
      - ApiKey: []
      summary: Delete a user
      tags:
      - users
    get:
      description: retrieve a specific user
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/httputil.HTTPResponse-model_UserWithScopes'
          description: OK
        "400":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/httputil.HTTPError'
          description: Bad Request
        "500":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/httputil.HTTPError'
          description: Internal Server Error
      security:
      - ApiKey: []
      summary: Get a user
      tags:
      - users
    put:
      description: updates a specific user
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/model.User'
        description: updated user info
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/httputil.HTTPResponse-model_User'
          description: OK
        "400":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/httputil.HTTPError'
          description: Bad Request
        "500":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/httputil.HTTPError'
          description: Internal Server Error
      security:
      - ApiKey: []
      summary: Update a user
      tags:
      - users
      x-codegen-request-body-name: data
  /admin/users/:id/scopes:
    put:
      description: update scopes for a specific user
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/controller.ScopeRequest'
        description: scopes to modify
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/httputil.HTTPResponse-model_UserWithScopes'
          description: OK
        "400":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/httputil.HTTPError'
          description: Bad Request
        "500":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/httputil.HTTPError'
          description: Internal Server Error
      summary: Update a user's scope
      tags:
      - users
      x-codegen-request-body-name: data
...
components:
  schemas:
...
    controller.ScopeRequest:
      properties:
        add:
          items:
            type: string
          type: array
        remove:
          items:
            type: string
          type: array
      type: object
    httputil.HTTPError:
      properties:
        code:
          example: 400
          type: integer
        message:
          example: status bad request
          type: string
      type: object
...
    httputil.HTTPResponse-array_model_User:
      example:
        data:
        - updated_at: 1
          name: name
          created_at: 0
          id: id
          deleted_at: 6
          email: email
        - updated_at: 1
          name: name
          created_at: 0
          id: id
          deleted_at: 6
          email: email
        message: message
      properties:
        data:
          items:
            $ref: '#/components/schemas/model.User'
          type: array
        message:
          type: string
      type: object
...
    httputil.HTTPResponse-model_User:
      example:
        data:
          updated_at: 1
          name: name
          created_at: 0
          id: id
          deleted_at: 6
          email: email
        message: message
      properties:
        data:
          $ref: '#/components/schemas/model.User'
        message:
          type: string
      type: object
    httputil.HTTPResponse-model_UserWithScopes:
      example:
        data:
          authentication_scopes:
          - authentication_scopes
          - authentication_scopes
          user:
            updated_at: 1
            name: name
            created_at: 0
            id: id
            deleted_at: 6
            email: email
        message: message
      properties:
        data:
          $ref: '#/components/schemas/model.UserWithScopes'
        message:
          type: string
      type: object
...
    model.User:
      example:
        updated_at: 1
        name: name
        created_at: 0
        id: id
        deleted_at: 6
        email: email
      properties:
        created_at:
          type: integer
        deleted_at:
          type: integer
        email:
          type: string
        id:
          type: string
        name:
          type: string
        updated_at:
          type: integer
      type: object
    model.UserWithScopes:
      example:
        authentication_scopes:
        - authentication_scopes
        - authentication_scopes
        user:
          updated_at: 1
          name: name
          created_at: 0
          id: id
          deleted_at: 6
          email: email
      properties:
        authentication_scopes:
          items:
            type: string
          type: array
        user:
          $ref: '#/components/schemas/model.User'
      type: object
  securitySchemes:
    ApiKey:
      in: header
      name: X-API-KEY
      type: apiKey
...